+++ src/main/webapp/WEB-INF/fragments/adminSidebar.jsp
@@ -68,12 +68,12 @@
               </a>
             </li>
             <li class="nav-item ">
-              <a class="nav-link" href="${pageContext.request.contextPath}/member/info">
+              <a class="nav-link" href="${pageContext.request.contextPath}/volt/pages/tables/bootstrap-tables.html">
                 <span class="sidebar-text">회원 조회</span>
               </a>
             </li>
             <li class="nav-item ">
-              <a class="nav-link" href="${pageContext.request.contextPath}/report/userReport">
+              <a class="nav-link" href="${pageContext.request.contextPath}/volt/pages/tables/bootstrap-tables.html">
                 <span class="sidebar-text">회원 신고 관리</span>
               </a>
             </li>
